split helper functions into different files
